Combine OpImplNoteParser and YamlJavadocBuilder
